+++ Versionfile
@@ -1,0 +1,2 @@
+"0.80.x" => { :branch => "master" }
+"0.30.x" => { :version => "1.0.1", :ref => '2bcca17f140fd4b8fab200eaf0d6636b2476ea4d'}
