+++ app/overrides/import_admin_tabs.rb
@@ -1,0 +1,5 @@
+Deface::Override.new(:virtual_path => "spree/layouts/admin",
+                     :name => "promo_admin_tabs",
+                     :insert_bottom => "[data-hook='admin_tabs'], #admin_tabs[data-hook]",
+                     :text => "<%= tab(:product_imports) %>",
+                     :disabled => false)
