+++ app/views/spree/user_mailer/product_import_results.text.erb
