+++ config/routes.rb
@@ -1,4 +1,4 @@
-Rails.application.routes.draw do
+Spree::Core::Engine.routes.prepend do
     namespace :admin do
       resources :product_imports, :only => [:index, :new, :create]
     end
