+++ db/migrate/20111118225054_rename_product_imports_table_to_spree_product_imports.rb
@@ -1,0 +1,5 @@
+class RenameProductImportsTableToSpreeProductImports < ActiveRecord::Migration
+  def change
+    rename_table :product_imports, :spree_product_imports
+	end
+end
