--- lib/import_products_hooks.rb
@@ -1,6 +1,0 @@
-class ImportProductsHooks < Spree::ThemeSupport::HookListener
-  # custom hooks go here
-  insert_after :admin_tabs do
-   %(<%= tab(:product_imports) %>)
-  end
-end
